+++ examples/stock.rs
@@ -10,7 +10,7 @@ fn parse_time(t: &str) -> Date<Local> {
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let data = get_data();
     let backend = BitMapBackend::new("examples/outputs/stock.png", (1024, 768));
-    let root: DrawingArea<_, _> = backend.into();
+    let root = backend.into_drawing_area();
     let font = Into::<FontDesc>::into("Arial").resize(50.0);
     root.fill(&RGBColor(255, 255, 255))?;